restructured requirements

+++ models/investment.py
@@ -7,4 +7,4 @@ class Investment(SqlAlchemyBase):
     title: str = sa.Column(sa.String, nullable=False, index=True)
     id: str = sa.Column(sa.String, primary_key=True)
     amount: int = sa.Column(sa.Integer, nullable=False)
-    coupon: int = sa.Column(sa.Float, nullable=False)
+    coupon: float = sa.Column(sa.Float, nullable=False)
